fix coding style

+++ algorithm/C++/0003-longest_substring_without_repeating.cpp
@@ -4,11 +4,11 @@ public:
         unordered_map<char, int> map;
         int j = 0;
         int maxNum = 0;
-        for(short i = 0; i < s.length(); ++i){
-            if(map.find(s[i]) != map.end()){
+        for (int i = 0; i < s.length(); ++i) {
+            if (map.find(s[i]) != map.end()) {
                 j = max(j, map[s[i]] + 1);
             }
-            if(maxNum < i - j + 1){
+            if (maxNum < i - j + 1) {
                 maxNum = i - j + 1;
             }
             map[s[i]] = i;
